+++ services/service_views/category.py
@@ -10,57 +10,10 @@ from authentication.decorators import validate_request
 from services.serializers.category_serializer import CategorySerializer, CategoryUpdateSerializer
 from services.serializers.provider_serializer import ProviderSerializer
 from services.service_models.category import Category
-from drf_spectacular.utils import extend_schema_view, extend_schema, OpenApiResponse
+from services.schemas.category_schemas import category_schema, category_provider_schema
 
 
-@extend_schema_view(
-    list=extend_schema(
-        summary="List Categories",
-        description="Retrieve a list of all categories.",
-        responses={
-            200: CategorySerializer(many=True)
-        },
-    ),
-    retrieve=extend_schema(
-        summary="Retrieve Category",
-        description="Retrieve details of a specific category by ID.",
-        responses={
-            200: CategorySerializer,
-        },
-    ),
-    create=extend_schema(
-        summary="Create Category",
-        description="Create a new category. Only accessible by super admins.",
-        request=CategorySerializer,
-        responses={
-            201: CategorySerializer,
-        },
-    ),
-    update=extend_schema(
-        summary="Update Category",
-        description="Update details of an existing category. Only accessible by super admins.",
-        request=CategoryUpdateSerializer,
-        responses={
-            200: CategoryUpdateSerializer,
-        },
-    ),
-    partial_update=extend_schema(
-        summary="Partial Update Category",
-        description="Partially update details of an existing category. Only accessible by super admins.",
-        request=CategoryUpdateSerializer,
-        responses={
-            200: CategoryUpdateSerializer,
-        },
-    ),
-    destroy=extend_schema(
-        summary="Delete Category",
-        description="Delete a category. Only accessible by super admins.",
-        responses={
-            204: OpenApiResponse(description="Category deleted successfully."),
-        },
-    ),
-)
-
+@category_schema
 class CategoryViewSet(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticated]
     queryset = Category.objects.all()
@@ -81,14 +34,7 @@ class CategoryViewSet(viewsets.ModelViewSet):
             return Response({"message": "An unexpected error occurred."}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
-    @extend_schema(
-        summary="List Providers for a Category",
-        description="Retrieve all providers associated with a specific category. "
-                    "This endpoint returns a list of providers that belong to the specified category.",
-        responses={
-            200: ProviderSerializer(many=True),
-        },
-    )
+    @category_provider_schema
     @action(detail=True, methods=['get'], permission_classes=[IsAuthenticated])
     def providers(self, request, pk=None):
         try:
